jarvis_ai_p
In this Project Jarvis is Digital Life Assistant which uses mainly  human  communication  means  such  Twitter,  instant message  and  voice  to  create  two  way  connections  between human and his apartment,  controlling lights and appliances, assist  in  cooking,  notify  him  of  breaking  news,  Facebook’s Notifications    and  many  more.  In  our  project  we  mainly use voice as communication means so the Jarvis is basically the Speech   recognition   application.   The   concept    of speech technology  really  encompasses  two  technologies:  Synthesizer and  recognizer.  A   speech  synthesizer  takes  as  input  and produces  an  audio  stream  as  output.  A  speech  recognizer on the  other hand does opposite. It takes an audio stream as input and thus turns it into text transcription. The  voice is a signal  of  infinite information. A direct analysis and synthesizing    the  complex  voice  signal  is  due  to  too  much information  contained in  the  signal.  Therefore  the   digital signal   processes   such   as Feature  Extraction   and  Feature Matching  are  introduced  to represent  the  voice  signal.